drbdmeta: fix apply-al for bitmap sizes > 4GiB
A long time ago, the maximum supported bitmap size was 512 MiB.
Nowadays, we can do Petabyte volumes and multiple peer bitmaps.

Using "unsigned int" for the size of the bitmap is just wrong.

Use uint64_t for .bm_bytes and .bits_set members.
Use size_t for extent sizes, to not 32bit truncate multiplication results.

Should we ever want to support platforms with size_t != off_t,
and a bitmap that is not completely in-core,
we need to double check all size_t declarations.

+++ user/shared/drbdmeta.h
@@ -98,8 +98,9 @@ struct format {
 
 	/* unused in 06 */
 	int md_index;
-	unsigned int bm_bytes;
-	unsigned int bits_set;	/* 32 bit should be enough. @4k ==> 16TB */
+
+	uint64_t bm_bytes;
+	uint64_t bits_set;
 	int bits_counted:1;
 	int update_lk_bdev:1;	/* need to update the last known bdev info? */
